+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Ability to style underlying WebView
+- Prop to handle errors on the WebView
+
 ### Changed
 ### Fixed